Linting. Needed some re-arranging this time.
